requirements/base.txt
# Base requirements for CyberHunter Security Portal
# Python 3.11+

# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Session==0.5.0
Flask-Cors==4.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0

# OAuth
Authlib==1.2.1
requests==2.31.0
requests-oauthlib==1.3.1

# Forms and Validation
WTForms==3.1.0
email-validator==2.0.0

# Security
cryptography==41.0.4
PyJWT==2.8.0
passlib==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
pyotp==2.9.0  # For TOTP/MFA
qrcode==7.4.2  # For MFA QR codes
python-secrets==22.1.0  # For secure token generation

# Email
imapclient==2.3.1
html2text==2020.1.16

# File Processing
python-magic==0.4.27
Pillow==10.0.1
pyclamd==0.4.0

# Background Tasks
celery==5.3.1
redis==4.6.0
flower==2.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
marshmallow==3.20.1
bleach==6.0.0

# Monitoring & Logging
python-json-logger==2.0.7
colorlog==6.7.0